Rename few functions to make those consistent
