+++ packages/prover/src/index.ts
@@ -1,2 +1,2 @@
 export * from "./interfaces.js";
-export {makeProvableProvider} from "./web3_provider.js";
+export {createVerifiedExecutionProvider} from "./web3_provider.js";
